Fix test_time test and ensure coverage
One of the test_time methods did not actually assert the exception message it was trying to detect as the assertion statement was contained within the context manager handling the exception. I've moved it out of the context so it actually runs.

I've also added a few `praga: no cover` comments for parts that were artifically lowering coverage of the test suite.