Always load doc and verification extensions
They used to only be loaded in "debug mode" because the main guild was
used to test the bot. However, we have since moved to using a separate
test guild so it's no longer a concern if these cogs get loaded.

It was confusing to some contributors as to why these cogs were not
being loaded since the debug mode isn't really documented anywhere.