HelpChannels: implement create_channel_queue
It returns a queue of dormant channels in random order. The queue will
be used to get the next available channel.

Using a random order is simpler than trying to sort by the timestamp
of the most recent message in each channel and this decision will only
"negatively" impact the system when the bot restarts or the extension
is reloaded. Ultimately, it just means in such events some dormant
channels may chosen to become active again sooner than expected.