HelpChannels: explicitly specify if a task should be cancelled
When rescheduling an idle channel, the task will only be cancelled if
the function was told the channel should currently have a task
scheduled. This means warnings for missing tasks will appear when they
should.

The previous approach of checking if a task exists was flawed because
it had no way to tell whether a task *should* exist. It assumed nothing
is wrong if a task doesn't exist.

Currently, the only case when a task shouldn't exist is when the
cog is initialised and channels from the bot's previous life are being
scheduled.