CI: mock the pipenv binary
The mock gets used by the flake8 pre-commit hook, which invokes flake8
via `pipenv run flake8`. It's normally useful to use pipenv here cause
it ensures flake8 is invoked within the context of the venv. However, in
CI, there is no venv - dependencies are installed directly to the
system site-packages. `pipenv run` does not work in such case because it
tries to create a new venv if one doesn't exist (it doesn't consider the
system interpreter to be a venv).

This workaround (okay, it's a hack) creates an executable shell script
which replaces the original pipenv binary. The shell script simply
ignores the first argument (i.e. ignores `run` in `pipenv run`) and
executes the rest of the arguments as a command. It essentially makes
`pipenv run flake8` equivalent to just having ran `flake8`. When
pre-commit executes pipenv, the aforementioned script is what will run.