Cog tests: add a function to get all commands
For tests, ideally creating instances of cogs should be avoided to avoid
extra code execution. This function was copied over from discord.py
because their function is not a static method, though it still works as
one. It was probably just a design decision on their part to not make it
static.