CI: install pipenv to user site
Some of pipenv's dependencies overlap with dependencies in the Pipfile.
When installing from the Pipfile, any dependencies already present in
the global site will not be installed again to the user site, and thus
will not be cached. Therefore, pipenv is installed to the user site to
ensure all dependencies get cached.

* Move PATH prepend step before pipenv invocation