ModLog: ignore update channel events for help channels
The edit causes two channel update events to dispatch simultaneously:
one for the channel topic changing and one for the category changing.
The ModLog cog currently doesn't support ignoring multiple events of
the same type for the same channel. Therefore, the ignore was hard coded
rather than using the typical ignore mechanism.

This is intended to be a temporary solution; it should be removed once
the ModLog is changed to support this situation.