Stop setting positions when moving help channels
Unfortunately, trying to set positions for the help channels during their move from one category to another does not work to well. It introduces a number of glitches and we haven't been able to reliably get a channel to go to a specific position either.

This commit simply removes any attempt to set a position and lets Discord handle it.