HelpChannels: handle errors when changing cooldown role
A user may leave the guild before their role can be changed. Sometimes,
there could also be role hierarchy issues or other network issues. It's
not productive to halt everything and just dump these as exceptions to
the loggers. The error handler provides a more graceful approach to
these exceptions.

* Add a wrapper function around `add_roles` & `remove_roles` which
  catches exceptions