Use `send_help` to invoke command help
After the refactoring of the help command, we need to use the built-in
method of calling the help command: `Context.send_help`. As an argument,
the qualified name (a string containing the full command path, including
parents) of the command can be passed.

Examples:

- await ctx.send_help("reminders edit")

This would send a help embed with information on `!reminders edit` to
the Context.

- await ctx.send_help(ctx.command.qualified_name)

This would extract the qualified name of the command, which is the full
command path, and send a help embed to Context.

- await ctx.send_help()

This will send the main "root" help embed to the Context.