Don't rely on HDEL ignoring keys for .pop
Previously we would try to .delete keys that did not exist if a default
was provided when calling .pop. This is okay to do (because HDEL will
just ignore any attempts to delete non-existing calls), but it does add
an additional pointless API call to Redis, so I've added some validation
as a small optimization.

This also adds a few additional lines of documentation as requested by
@SebastiaanZ in their review.