DRY approach to typestring prefix resolution
Thanks to @kwzrd for this idea, basically we're making a constant with
the typestring prefixes and iterating that in all our converters.

These converter functions will also now raise TypeErrors if we try
to convert something that isn't in this constants list.

I've also added a new test that tests this functionality.