Make .items return ItemsView instead of AsyncIter
There really was no compelling reason why this method should return an
AsyncIterator or than that `async for items in cache.items()` has nice
readability, but there were a few concerns. One is a concern about race
conditions raised by @SebastiaanZ, and @MarkKoz raised a concern that it
was misleading to have an AsyncIterator that only "pretended" to be
lazy.

To address these concerns, I've refactored it to return a regular
ItemsView instead.

I also improved the docstring, and fixed the relevant tests.