Add .increment and .decrement methods.
Sometimes, we just want to store a counter in the cache. In this case,
it is convenient to have a single method that will allow us to increment
or decrement this counter.

These methods allow you to decrement or increment floats and integers by
an specified amount. By default, it'll increment or decrement by 1.

Since this involves several API requests, we create an asyncio.Lock so
that we don't end up with race conditions.