Add an option to use fakeredis in Bot.
Without this option, all contributors would need to set up a Redis
server in order to run the bot. But with use_fakeredis set to True, this
is no longer necessary because it will just set up a fakeredis redis
pool instead of trying to contact an actual server.

This is more than good enough for most local testing purposes, since
data persistence across restarts isn't really relevant for them.

This also means we need to move fakeredis into our real dependency list
instead of having it as a dev dependency, so there's a minor change for
that as well.

I also made a small kaizen change to sort all the dependencies in the
Pipfile alphabetically.