Add support for Union type annotations for constants
Note that `Optional[x]` is just an alias for `Union[None, x]` so this
effectively supports `Optional` too.

This was especially troublesome because the redis password must be
unset/None in order to avoid authentication, but the test would complain
that `None` isn't a `str`. Setting to an empty string would pass the
test but then make redis authenticate and fail.