Token remover: use regex groups and pass the token as a NamedTuple
It felt redundant to be splitting the token in two different functions
when regex could take care of this from the outset. '

A NamedTuple was created to house the token. This is nicer than passing
an re.Match object, because it's clearer which attributes are available.
Even if the regex used named groups, it wouldn't be as obvious which
group names exist.

Without the split, `is_maybe_token` is dwindled down to a redundant
function. Therefore, it's been removed.