Refactor the in_whitelist deco to a check.
We're moving the actual predicate into the `utils.checks` folder, just
like we're doing with most of the other decorators. This is to allow us
the flexibility to use it as a pure check, not only as a decorator.

This commit doesn't actually change any functionality, just moves it
around.