Add a mod_log.ignore_all context manager.
This new context manager makes it easier to make the mod_log ignore
actions like message deletions. The only existing method is the
`ignore()` method, which requires that you pass all the messages you
want to ignore into it.

This one just ignores everything inside its scope.

This isn't the DRYest approach, but it's low-cost and improves the
readability of clean.py quite a bit. Ideally we should go through and
give modlog a proper cleanup, because it's kinda ugly right now.