Fix a bug with invocation deletion.
This command was written to support only a single channel, and with the
move to multi-channel purges, we need to rethink the way the invocation
deletion happens. We may be invoking this command from a completely
different channel, so we can't necessarily look inside the channels
we're targeting for the invocation.

So, we're solving this by just deleting the invocation by using
ctx.message. We do this before we start iterating message history, and
then we only need to iterate the number of messages that was passed into
the command.

A much cleaner approach, which solves the bug reported and identified by
@MarkKoz.