Incidents: implement `process_event` coroutine
This contains the main logic for handling reactions and glues all the
helpers together.

Unfortunately, gracefully handling everything that can go wrong in
the process requires quite a lot of code ~ but, at least to me, it
seems like this all should now be fairly safe.

The idea to await the message delete event before releasing the lock
was conceived by Ves, while Mark helped me refine it.

Co-authored-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>
Co-authored-by: MarkKoz <kozlovmark@gmail.com>