ModLog: fix AttributeError in on_member_update
`iterable_item_removed` and `iterable_item_added` lack `new_value` and
`old_value`. Instead, they just contain the actual value added or
removed. The code was incorrectly trying to access old and new values
for the iterable changes.

The iterable changes are only useful for the role diff, but they aren't
even needed for that. The role diff calculation has been refactored to
always get the diff rather than doing it only if it sees there has
been a change to the `_roles` attribute.

To be clear, `_roles` only has IDs, which is why its diff isn't that
useful anyway. To use it, the code would have to get the Role objects,
which is basically what the `roles` property already does. `_cs_roles`
seems to be some Role object cache, but its reliability is unclear.