Allow role pings in Syncers and help_channels.py
Now that we're running Discord 1.4.0a, we need to explicitely allow all
the role mentions for sends that don't use ping one of the globally
whitelisted role pings, which are Moderators, Admins and Owners.

We were pinging roles other than Mods+ in exactly two cases:
- Inside the Syncers, whenever we ask for sync confirmation (if the
  number of roles or users to sync is unusually high)
- In the help_channels.py system, whenever we max out help channels and
  are unable to create more.

This commit addresses both of these.

GitHub #1038
https://github.com/python-discord/bot/issues/1038