Incidents: link `proxy_url` if attachment fails to download
Suggested by Mark during review. If the download fails, we fallback
on showing an informative message, which will link the attachment
cdn link.

The attachment-handling logic was moved from the `archive` coroutine
into `make_embed`, which now also returns the file, if available.
In the end, this appears to be the smoothest approach.

Co-authored-by: MarkKoz <kozlovmark@gmail.com>