Remove caching of last_dm_user.
If you're typing up a reply and the bot gets another DM while you're
typing, you might accidentally send your reply to the wrong person.

This could happen even if you're very attentive, because it might be a
matter of milliseconds. The complexity to prevent this isn't worth the
convenience of the feature, and it's nice to get rid of the caching as
well, so I've decided to just make .reply require a user for every
reply.

https://github.com/python-discord/bot/issues/1041