Cache AllowDenyList data at bot startup.
We shouldn't be making an API call for every single message posted, so
what we're gonna do is cache the data in the Bot, and then update the
cache whenever we make changes to it via our new AllowDenyList cog.

Since this cog will be the only way to make changes to this, this level
of lazy caching should be enough to always keep the cache up to date.