Kaizen: Make error_handler.py more embeddy.
Currently, some types of errors are returning plain strings that repeat
the input (which can be exploited to deliver stuff like mentions), and
others are returning generic messages that don't give any exception
information.

This commit unifies our approach around putting as much information as
we can (including the exception message), but always putting it inside
an embed, so that stuff like pings will not fire.

This, combined with the 1.4.0a `allowed_mentions` functionality, seems
like a reasonable compromise between security and usability.