Validation of guild invites.
We will now validate and convert any standard discord server invite to a
guild ID, and automatically add the name of the server as a comment.

This will ensure that the list of whitelisted guild IDs will be readable
and nice.

This also makes minor changes to list output aesthetics.