Add function for finding tags until a matching tag
This will allow flexibility in the future when collecting tags
for the description and signature of symbols.
The base is a function which accepts a callable which is called and
iterated over, but 3 names with a partial function that has the callable
supplied are provided to keep the outside interface neater.