Simplify module parsing method.
Instead of returning None and multiple values, the method now
only returns the string of the description.
Previously the parsing returned None and quit
when appropriate tags for shortening the description
were not found, but the new implementation simply defaults to the
provided start tag if a better alternative is not found.