Properly parse labels add fallback for non dt tags
Labels point to tags that aren't in description lists, like modules
or doc symbols which we already handle.
If by chance we get a symbol that we don't have in the group for
general parsing and which isn't a dt tag, log it and don't attempt to
parse signature and use general description parsing instead of parsing a
dd tag.