Don't ping everyone when tripping filter in DMs.
We don't need a ping in #mod-alerts whenever someone is tripping a
filter (like invites or bad language) in a DM to the bot. We can still
send an embed, so that we can action it, but there is no urgent need to
respond if it's just a direct message to the bot.

This is particularly true now that we have #dm-log.