Dynamically amend types to filterlist docstrings.
We want the !help invocations to give you all the information you need
in order to use the command. That also means we need to provide the
valid filterlist types, which are subject to change.

So, we fetch the valid ones from the API and then dynamically insert
them into the docstrings.