Silence: revoke permissions to add reactions
No longer assume default values for the overwrites which are modified.
Save and restore previous values `add_reactions` and `send_messages`
via redis.

When unsilencing, check if a channel is silenced via the redis cache
rather than the channel's current overwrites to ensure the task is
cancelled even if overwrites were manually edited.